+++ widgets/events/components/list/index.ts
@@ -1,0 +1,1 @@
+export { List, type ListItem } from './list';
