+++ widgets/events/style.module.css
@@ -1,16 +1,15 @@
-
 .locationPicker {
-	margin-block-end: 12px;
-	max-width: 400px;
+	margin-block-end: var(--wpds-dimension-padding-md);
+	max-width: calc(var(--wpds-dimension-base) * 100);
 }
 
 .locationInput {
 	flex: 1;
-	min-inline-size: 120px;
+	min-inline-size: calc(var(--wpds-dimension-base) * 30);
 }
 
 .eventNone {
-	padding-block: 8px;
+	padding-block: var(--wpds-dimension-padding-sm);
 	color: var(--wpds-color-fg-content-neutral-weak);
 }
 
@@ -20,32 +19,12 @@
 }
 
 .footer {
-	padding: 12px 0 0;
-	margin: 12px 0 0;
-	border-block-start: 1px solid var(--wpds-color-stroke-surface-neutral);
+	padding-block-start: var(--wpds-dimension-padding-md);
+	margin-block-start: var(--wpds-dimension-padding-md);
+	border-block-start: var(--wpds-border-width-sm) solid
+		var(--wpds-color-stroke-surface-neutral-weak);
 }
 
 .footerLinks {
 	margin-inline-start: auto;
-}
-
-.listItem {
-	inline-size: 100%;
-}
-
-.listItemIcon {
-	inline-size: 36px;
-	block-size: 36px;
-	display: inline-flex;
-	align-items: center;
-	justify-content: center;
-	border-radius: 8px;
-	background: var(--wpds-color-bg-track-neutral-weak);
-	color: var(--wpds-color-fg-content-neutral-weak);
-	flex-shrink: 0;
-}
-
-.listItemContent {
-	flex: 1;
-	min-inline-size: 0;
 }
